fix(plugins): keep workspace identity out of doctor state imports
