+++ src/memory-host-sdk/event-store.ts
@@ -1,6 +1,6 @@
 import { createHash } from "node:crypto";
 import { z } from "zod";
-import { resolveWorkspaceStateIdentity } from "../agents/workspace-state-store.js";
+import { resolveWorkspaceStateIdentity } from "../agents/workspace-state-identity.js";
 import {
   pluginStateEntriesInKeyRange,
   registerPluginStateSyncSequencedJournalEntry,
